complex_calc/calc/calculator_test.go
package calc

import "testing"

func TestCalculate(t *testing.T) {
	type args struct {
		exp string
	}
	tests := []struct {
		name    string
		args    args
		want    string
		wantErr bool
	}{
		// TODO: Add test cases.
		{
			name: "test1",
			args: args{
				exp: "10 + 9 - 3",
			},
			want:    "16",
			wantErr: false,
		},
		{
			name: "test1",
			args: args{
				exp: "10 + 9 / 0",
			},
			want:    "-999",
			wantErr: true,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			got, err := Calculate(tt.args.exp)
			if (err != nil) != tt.wantErr {
				t.Errorf("Calculate() error = %v, wantErr %v", err, tt.wantErr)
				return
			}
			if got != tt.want {
				t.Errorf("Calculate() = %v, want %v", got, tt.want)
			}
		})
	}
}
